Remove app.routes.ts
All routes handled in a single file
