+++ src/app/routing/app-routing.module.ts
@@ -1,8 +1,26 @@
 import { NgModule } from '@angular/core';
-import { RouterModule } from '@angular/router';
+import { RouterModule, Routes } from '@angular/router';
 
-import { ROUTES } from './app.routes';
+import { LoggedInGuard, UserComponent } from 'ngx-auth-firebaseui';
 
+import { HomeComponent } from './home/home.component';
+import { BookComponent } from '../book/book/book.component';
+import { SigninComponent } from './signin/signin.component';
+import { Error404Component } from './error-404/error-404.component';
+
+export const ROUTES: Routes = [
+  { path: '', redirectTo: '/signin', pathMatch: 'full' },
+  {
+    path: 'home', component: HomeComponent, canActivate: [LoggedInGuard],
+    children: [
+      { path: ':bookUid', component: BookComponent }
+    ]
+  },
+  { path: 'signin', component: SigninComponent },
+  { path: 'user', component: UserComponent, canActivate: [LoggedInGuard] }, // ngx-auth-firebaseui-user
+  { path: '**', component: Error404Component },
+
+];
 
 @NgModule({
   declarations: [],
